added some code to test CRUD operations
added some code to test CRUD operations

+++ src/main/java/com/color/finalsprints4/service/UserService.java
@@ -1,12 +1,12 @@
 package com.color.finalsprints4.service;
 
-import com.color.finalsprints4.model.Color;
 import com.color.finalsprints4.model.User;
 import com.color.finalsprints4.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -23,11 +23,26 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
-
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("User not found with id: " + id); // Handle this appropriately
+        }
     }
 
     public User updateUser(Long id, User updatedUser) {
-        return updatedUser;
+        Optional<User> existingUser = userRepository.findById(id);
+        if (existingUser.isPresent()) {
+            User user = existingUser.get();
+            // Update fields
+            user.setFirstName(updatedUser.getFirstName());
+            user.setLastName(updatedUser.getLastName());
+            user.setEmail(updatedUser.getEmail());
+            user.setSelectedColor(updatedUser.getSelectedColor());
+            return userRepository.save(user);
+        } else {
+            throw new RuntimeException("User not found with id: " + id);
+        }
     }
 }
+
